+++ CHANGELOG.md
@@ -31,6 +31,9 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 - Refactoring stories to simpler setup.
 - Linting fixes.
 
+- Added footer component
+- Added copyright component
+
 ## [3.0.11] - 2020-01-29
 
 ## Changed